Better error messages for invalid types and composed properties.
Error messages for invalid properties now include the required type (a list of possible types for MultiTypeProcessor) and the actually given type.

Error messages for composed values now include how many of the composition elements must be matched and how many are matched. Additionally appends a list of all composition elements to the exception message which shows the elements which were matched and those who failed (only available if the error collection mode is enabled). Failing elements additionally include the exact reason for failing (the collected exception messages from the nested validation)