Fix regression introduced by #68
Validators which render a separate method and are transferred to a parent schema will render the method to the origin schema. Consequently, a call to the validator in the parent schema will cause a "call to undefined method" error. To fix the issue the validator scope gets updated at render time to make sure the method is rendered to the correct class.